refactor: eliminate relative imports, create llm/models package, move inline imports to top-of-file
