+++ src/truthfulness_evaluator/__init__.py
@@ -9,27 +9,44 @@ except PackageNotFoundError:
 
 # Legacy graph constructor
 # Protocols
-from .core.protocols import ClaimExtractor, ClaimVerifier, EvidenceGatherer, ReportFormatter
+from truthfulness_evaluator.core.protocols import (
+    ClaimExtractor,
+    ClaimVerifier,
+    EvidenceGatherer,
+    ReportFormatter,
+)
 
 # Workflow infrastructure
-from .llm.workflows import WorkflowConfig, WorkflowRegistry
-from .llm.workflows.graph import create_truthfulness_graph
-from .llm.workflows.presets import register_builtin_presets
+from truthfulness_evaluator.llm.workflows import WorkflowConfig, WorkflowRegistry
+from truthfulness_evaluator.llm.workflows.graph import create_truthfulness_graph
+from truthfulness_evaluator.llm.workflows.presets import register_builtin_presets
 
 # Domain models
-from .models import Claim, TruthfulnessReport, VerificationResult
+from truthfulness_evaluator.models import Claim, TruthfulnessReport, VerificationResult
 
 # Strategy adapters - Extractors
-from .strategies.extractors import SimpleExtractor, TripletExtractor
+from truthfulness_evaluator.strategies.extractors import SimpleExtractor, TripletExtractor
 
 # Strategy adapters - Formatters
-from .strategies.formatters import HtmlFormatter, JsonFormatter, MarkdownFormatter
+from truthfulness_evaluator.strategies.formatters import (
+    HtmlFormatter,
+    JsonFormatter,
+    MarkdownFormatter,
+)
 
 # Strategy adapters - Gatherers
-from .strategies.gatherers import CompositeGatherer, FilesystemGatherer, WebSearchGatherer
+from truthfulness_evaluator.strategies.gatherers import (
+    CompositeGatherer,
+    FilesystemGatherer,
+    WebSearchGatherer,
+)
 
 # Strategy adapters - Verifiers
-from .strategies.verifiers import ConsensusVerifier, InternalVerifier, SingleModelVerifier
+from truthfulness_evaluator.strategies.verifiers import (
+    ConsensusVerifier,
+    InternalVerifier,
+    SingleModelVerifier,
+)
 
 __all__ = [
     "__version__",
