+++ src/truthfulness_evaluator/evidence/agent.py
@@ -6,9 +6,9 @@ from typing import Any
 
 from langgraph.prebuilt import create_react_agent
 
-from ..core.logging_config import get_logger
-from ..llm import create_chat_model
-from .tools.filesystem import get_filesystem_tools
+from truthfulness_evaluator.core.logging_config import get_logger
+from truthfulness_evaluator.evidence.tools.filesystem import get_filesystem_tools
+from truthfulness_evaluator.llm import create_chat_model
 
 logger = get_logger()
