+++ src/truthfulness_evaluator/llm/models/__init__.py
@@ -1,0 +1,21 @@
+from .evidence import EvidenceAnalysisItem, EvidenceAnalysisOutput
+from .extraction import (
+    ClaimExtractionOutput,
+    ExtractedClaim,
+    KnowledgeTriplet,
+    TripletExtractionOutput,
+)
+from .internal_verification import ClaimClassification, InternalVerificationOutput
+from .verification import VerificationOutput
+
+__all__ = [
+    "ExtractedClaim",
+    "ClaimExtractionOutput",
+    "KnowledgeTriplet",
+    "TripletExtractionOutput",
+    "VerificationOutput",
+    "EvidenceAnalysisItem",
+    "EvidenceAnalysisOutput",
+    "ClaimClassification",
+    "InternalVerificationOutput",
+]
