+++ src/truthfulness_evaluator/strategies/formatters/markdown.py
@@ -1,6 +1,7 @@
 """Markdown report formatter."""
 
-from ...models import TruthfulnessReport
+from truthfulness_evaluator.models import TruthfulnessReport
+from truthfulness_evaluator.reporting.generator import ReportGenerator
 
 
 class MarkdownFormatter:
@@ -8,8 +9,6 @@ class MarkdownFormatter:
 
     def format(self, report: TruthfulnessReport) -> str:
         """Format a truthfulness report as Markdown."""
-        from ...reporting.generator import ReportGenerator
-
         gen = ReportGenerator(report)
         return gen.to_markdown()
