+++ src/truthfulness_evaluator/strategies/verifiers/internal.py
@@ -1,8 +1,11 @@
 """Internal/codebase verification adapter."""
 
-from ...core.logging_config import get_logger
-from ...llm.chains.internal_verification import ClaimClassifier, InternalVerificationChain
-from ...models import Claim, Evidence, VerificationResult
+from truthfulness_evaluator.core.logging_config import get_logger
+from truthfulness_evaluator.llm.chains.internal_verification import (
+    ClaimClassifier,
+    InternalVerificationChain,
+)
+from truthfulness_evaluator.models import Claim, Evidence, VerificationResult
 
 logger = get_logger()
